+++ src/modules/layerTree/components/LayerTree.vue
@@ -16,9 +16,21 @@ export default {
         LayerTreeNode
     },
     computed: {
-        ...mapGetters(["allLayerConfigsStructured", "showLayerAddButton"]),
+        ...mapGetters(["allLayerConfigsStructured", "showLayerAddButton", "portalConfig"]),
         ...mapGetters("Modules/LayerTree", ["menuSide"]),
-        ...mapGetters("Modules/LayerSelection", {layerSelectionType: "type", layerSelectionName: "name"})
+        ...mapGetters("Modules/LayerSelection", {layerSelectionType: "type", layerSelectionName: "name"}),
+
+        /**
+         * Returns the title of button.
+         * @returns {String} the button title.
+         */
+        title () {
+            if (typeof this.portalConfig?.tree?.addLayerButton?.buttonTitle === "string" && this.portalConfig?.tree?.addLayerButton?.buttonTitle !== "") {
+                return this.portalConfig?.tree?.addLayerButton?.buttonTitle;
+            }
+
+            return "common:modules.layerTree.addLayer";
+        }
     },
     methods: {
         ...mapActions("Modules/LayerSelection", ["navigateForward"]),
@@ -60,9 +72,9 @@ export default {
         >
             <ElevatedButton
                 id="add-layer-btn"
-                aria-label="$t('common:modules.layerTree.addLayer')"
+                :aria-label="$t(title)"
                 :interaction="showLayerSelection"
-                :text="$t('common:modules.layerTree.addLayer')"
+                :text="$t(title)"
                 :icon="'bi-plus-circle'"
             />
         </div>