+++ src/modules/layerTree/tests/unit/components/LayerTree.spec.js
@@ -305,4 +305,32 @@ describe("src/modules/layerTree/components/LayerTree.vue", () => {
         expect(spy.calledOnce).to.be.true;
     });
 
+    it("with layer button - shows the default title", () => {
+        subjectDataLayers = [];
+        layersBG = [];
+        addLayerButton = {
+            active: true
+        };
+        wrapper = shallowMount(LayerTreeComponent, {
+            global: {
+                plugins: [store]
+            }});
+
+        expect(wrapper.find("#add-layer-btn").attributes().text).to.be.equals("common:modules.layerTree.addLayer");
+    });
+
+    it("with layer button - shows the configurable title", () => {
+        subjectDataLayers = [];
+        layersBG = [];
+        addLayerButton = {
+            active: true,
+            "buttonTitle": "test"
+        };
+        wrapper = shallowMount(LayerTreeComponent, {
+            global: {
+                plugins: [store]
+            }});
+
+        expect(wrapper.find("#add-layer-btn").attributes().text).to.be.equals("test");
+    });
 });