+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Modeler3D: Modeler3D was added
 - LayerTree:
     - In SubMenu added button to open contact form with layer specific parameters to be used in email.
+    - Adds parameter buttonTitle in addLayerButton to enable to configure the button title.
 
 ### Changed
 - LayerSelection
@@ -63,7 +64,7 @@
     - Added scheme request function in the api folder for the getOAFFeature file. This can be used for an easy and fast way to gather unique values.
 - ViewPoint:
     - Migrated from dev to dev_vue.
-- Matomo: Added Support for analytics via matomo.  
+- Matomo: Added Support for analytics via matomo.
 - Migrator: supports ids with suffix.
 - InputText:
     - Added change event handler property and disabled property.